pyproject.toml
[project]
name = "cheese3d"
version = "0.1.0"
description = "Sensitive detection and analysis of whole-face movement in mice."
authors = [{ name = "Hou Lab CSHL", email = "the.hou.lab@gmail.com" }]
readme = "README.md"
requires-python = "==3.10.*"
dependencies = [
    "ipython",
    "ipykernel",
    "jupyter",
    "protobuf < 3.20",
    "tables == 3.7.*",
    "pandas == 2.1.*",
    # "numpy == 1.23.2",
    # "numpy == 1.26.1",
    "configobj == 5.0.8",
    "hydra-core",
    "joblib",
    "torch <= 1.12.0",
    "napari-deeplabcut @ https://github.com/Hou-Lab-CSHL/napari-deeplabcut.git",
    "aniposelib @ https://github.com/Hou-Lab-CSHL/aniposelib.git#jax",
    "anipose @ https://github.com/Hou-Lab-CSHL/anipose.git",
    "opencv-contrib-python == 4.7.*",
    "opencv-python == 4.7.*",
    "opencv-python-headless == 4.7.*",
    # "vtk == 9.3.*",
    # "mayavi == 4.8.2",
    "av == 9.2.0",
    "pims == 0.6.*",
    "seaborn",
    "open-ephys-python-tools",
]

[tool.pixi.dependencies]
c-compiler = "*"
cxx-compiler = "*"
ffmpeg = "==4.3.2"

[tool.pixi.target.linux-64.dependencies]
cudatoolkit = "11.7.*"
cudnn = "8.*"

[tool.pixi.target.osx-arm64.dependencies]
tensorflow-deps = "*"
hdf5 = "*"
pytables = "*"

[tool.pixi.target.linux-64.pypi-dependencies]
blosc2 = "*"
tensorflow = "==2.10.*"
jax = { version = ">=0.4, <0.5", extras = ["cuda12_pip"] }
deeplabcut = { git = "https://github.com/Hou-Lab-CSHL/DeepLabCut.git", extras = ["gui", "tf"] }

[tool.pixi.target.osx-arm64.pypi-dependencies]
tensorflow-macos = "==2.10.*"
tensorflow-metal = "==0.6.*"
jax = { version = ">=0.4, <0.5", extras = ["cpu"] }
deeplabcut = { git = "https://github.com/Hou-Lab-CSHL/DeepLabCut.git", extras = ["gui", "apple_mchips"] }

[tool.pixi.project]
channels = ["conda-forge", "nvidia", "apple"]
platforms = ["linux-64", "osx-arm64"]

[tool.pixi.pypi-dependencies]
cheese3d = { path = ".", editable = true }

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.hatch.metadata]
allow-direct-references = true
